Enhance syscall error handling and logging
- Updated `SyscallError` enum to include additional error variants, improving alignment with standard error codes.
- Implemented `fmt::Display` for `SyscallError` to provide user-friendly error messages.
- Refactored error handling in syscall functions to utilize the new `SyscallError` variants, enhancing clarity and consistency.
- Improved logging of syscall errors by integrating error names and descriptions for better debugging.
- Updated syscall wrappers in `strate-fs-ext4` and `strate-net` to utilize the new error handling mechanisms.

+++ workspace/components/strate-net/src/main.rs
@@ -87,7 +87,7 @@ const NANOS_PER_MICRO: u64 = 1_000;
 
 fn now_instant() -> Instant {
     match clock_gettime_ns() {
-        Ok(ns) => Instant::from_micros(ns / NANOS_PER_MICRO),
+        Ok(ns) => Instant::from_micros((ns / NANOS_PER_MICRO) as i64),
         Err(_) => Instant::from_micros(0),
     }
 }
